+++ lib/Email/Service/MailgunHelper.php
@@ -3,126 +3,119 @@
 namespace Ridibooks\Platform\Common\Email\Service;
 
 use Mailgun\Exception\HttpClientException;
+use Mailgun\HttpClient\HttpClientConfigurator;
+use Mailgun\Hydrator\Hydrator;
 use Mailgun\Mailgun;
+use Mailgun\Model\EmailValidation\ValidateResponse;
+use Mailgun\Model\Message\SendResponse;
 use Ridibooks\Platform\Common\Email\Constant\EmailContentTypeConst;
-use Ridibooks\Platform\Common\Security\Encryption;
 
-/**
- * 환경변수 의존성 존재함
- * \Config::$MAILGUN_API_KEY
- * \Config::$MAILGUN_SEND_DOMAIN
- * \Config::$MAILGUN_PUBLIC_ID
- * \Config::$MAILGUN_PUBLIC_KEY
- * \Config::$MAILGUN_EMAIL_VALIDATE_API_URL
- * \Config::$MAILGUN_SECRET_KEY_OPT_OUT
- */
 class MailgunHelper
 {
-    private const RESPONSE_HTTP_OK = 200;
-    /** @var Mailgun */
-    private $mailgun;
+    /** @var HttpClientConfigurator */
+    private $configurator;
+    /** @var Hydrator */
+    private $hydrator;
 
-    public function __construct(string $api_key)
+    /** @var string */
+    private $send_domain;
+    /** @var bool */
+    private $is_testmode;
+
+    /** @var self */
+    private static $instance;
+
+    public static function getInstance(): self
     {
-        $this->mailgun = new Mailgun($api_key);
+        if (self::$instance === null) {
+            self::$instance = new self();
+        }
+
+        return self::$instance;
     }
 
-    public function getBounces(int $skip = 0, int $limit = 1)
+    private function __construct()
     {
-        $result = $this->mailgun->get(\Config::$MAILGUN_SEND_DOMAIN . "/bounces", ['skip' => $skip, 'limit' => $limit]);
+        $this->configurator = new HttpClientConfigurator();
+    }
 
-        return $result->http_response_body;
+    public function setApiKey(string $api_key): self
+    {
+        $this->configurator->setApiKey($api_key);
+
+        return $this;
+    }
+
+    public function setSendDomain(string $send_domain): self
+    {
+        $this->send_domain = $send_domain;
+
+        return $this;
+    }
+
+    public function setTestMode(bool $is_testmode, ?string $test_url = ''): self
+    {
+        if ($is_testmode && empty($test_url)) {
+            throw new \InvalidArgumentException('Invalid test_url');
+        }
+
+        $this->is_testmode = $is_testmode;
+        $this->configurator->setDebug($this->is_testmode);
+        if ($is_testmode) {
+            $this->configurator->setEndpoint($test_url);
+        }
+
+        return $this;
+    }
+
+    public function setHydrator(Hydrator $hydrator): self
+    {
+        $this->hydrator = $hydrator;
+
+        return $this;
     }
 
     /**
-     * @param string        $from
-     * @param string[]      $to
-     * @param string        $subject
-     * @param string        $content
+     * @param string   $from
+     * @param string[] $to
+     * @param string   $subject
+     * @param string   $content
      * @param string[] $cc
      * @param string[] $bcc
-     * @param string        $content_type
-     * @param array         $attachments
-     * @param bool          $is_testmode
+     * @param string   $content_type
+     * @param array    $attachments [filename => filepath]
      *
-     * @return \stdClass
-     * @throws \Mailgun\Messages\Exceptions\MissingRequiredMIMEParameters
+     * @return bool
      */
     public function send(
         string $from,
         array $to,
         string $subject,
         string $content,
+        string $content_type = EmailContentTypeConst::TEXT_HTML,
         array $cc = [],
         array $bcc = [],
-        string $content_type = EmailContentTypeConst::TEXT_HTML,
-        array $attachments = [],
-        bool $is_testmode = false
-    ) {
-        $message = self::prepareCommonParameters(
-            $from,
-            $to,
-            $subject,
-            $content,
-            $content_type,
-            $cc,
-            $bcc,
-            $is_testmode
-        );
-        $message['content-type'] = $content_type;
+        array $attachments = []
+    ): bool {
+        $this->assert();
+
+        $parameters = $this->prepareParameters($from, $to, $subject, $content, $content_type, $cc, $bcc);
 
         $formatted_attachments = [];
         foreach ($attachments as $name => $path) {
-            $formatted_attachments[] = self::generateMailgunAttachmentFormat($name, $path);
+            $formatted_attachments[] = $this->generateMailgunAttachmentFormat($name, $path);
         }
 
-        $files = [];
         if (!empty($formatted_attachments)) {
-            $files['attachment'] = $formatted_attachments;
+            $parameters['attachment'] = $formatted_attachments;
         }
-
-        return $this->mailgun->sendMessage(\Config::$MAILGUN_SEND_DOMAIN, $message, $files);
-    }
-
-    /**
-     * @param string        $from
-     * @param string[]      $to
-     * @param string        $subject
-     * @param string        $content
-     * @param string        $content_type
-     * @param string[] $cc
-     * @param string[] $bcc
-     * @param bool          $is_testmode
-     *
-     * @return bool
-     */
-    public static function sendUsingV3(
-        string $from,
-        array $to,
-        string $subject,
-        string $content,
-        string $content_type = EmailContentTypeConst::TEXT_HTML,
-        array $cc = [],
-        array $bcc = [],
-        bool $is_testmode = false
-    ) {
-        $mg = Mailgun::create(\Config::$MAILGUN_API_KEY);
-
-        $parameters = self::prepareCommonParameters(
-            $from,
-            $to,
-            $subject,
-            $content,
-            $content_type,
-            $cc,
-            $bcc,
-            $is_testmode
-        );
 
         try {
-            $mg->messages()->send(\Config::$MAILGUN_SEND_DOMAIN, $parameters);
+            $mailgun = new Mailgun($this->configurator, $this->hydrator);
+            /** @var SendResponse $result */
+            $result = $mailgun->messages()->send($this->send_domain, $parameters);
 
-            return true;
+            return !empty($result->getId());
         } catch (HttpClientException $e) {
             trigger_error($e->getMessage());
 
@@ -131,32 +124,29 @@ class MailgunHelper
     }
 
     /**
-     * @param string        $from
-     * @param string[]      $to
-     * @param string        $subject
-     * @param string        $content
+     * @param string   $from
+     * @param string[] $to
+     * @param string   $subject
+     * @param string   $content
      * @param string[] $cc
      * @param string[] $bcc
-     * @param string        $content_type
-     * @param bool          $is_testmode
+     * @param string   $content_type
      *
      * @return array
      */
-    private static function prepareCommonParameters(
+    private function prepareParameters(
         string $from,
         array $to,
         string $subject,
         string $content,
         string $content_type = EmailContentTypeConst::TEXT_HTML,
         array $cc = [],
-        array $bcc = [],
-        bool $is_testmode = false
+        array $bcc = []
     ): array {
         $parameters = [
             'from' => $from,
             'to' => implode(',', $to),
             'subject' => $subject,
-            'recipient-variables' => self::createRecipientVariables($to),
         ];
 
         if ($content_type === EmailContentTypeConst::TEXT_HTML) {
@@ -173,99 +163,40 @@ class MailgunHelper
             $parameters['bcc'] = $bcc;
         }
 
-        if ($is_testmode) {
-            $parameters['o:testmode'] = 'true';
+        if ($this->is_testmode) {
+            $parameters['o:testmode'] = 'yes';
         }
 
         return $parameters;
     }
 
-    /**
-     * @param \stdClass $results
-     *
-     * @return bool
-     */
-    public static function isSuccess($results): bool
+    public function isVaildEmail(string $email): bool
     {
-        if ($results->http_response_code != self::RESPONSE_HTTP_OK) {
-            trigger_error('[EMAIL][Mailgun] Error : ' . $results->http_response_code);
+        $this->assert();
 
+        try {
+            $mailgun = new Mailgun($this->configurator, $this->hydrator);
+            /** @var ValidateResponse $result */
+            $result = $mailgun->emailValidation()->validate($email);
+
+            return $result->isValid();
+        } catch (\Exception $e) {
             return false;
         }
-
-        return true;
     }
 
-    /**
-     * @param string $email
-     *
-     * @return bool
-     */
-    public static function isVaildEmail($email)
-    {
-        $url = \Config::$MAILGUN_EMAIL_VALIDATE_API_URL;
-        $querystring = 'address=' . urlencode($email);
-
-        $user_id = \Config::$MAILGUN_PUBLIC_ID;
-        $user_password = \Config::$MAILGUN_PUBLIC_KEY;
-
-        $ch = curl_init();
-        curl_setopt($ch, CURLOPT_URL, "{$url}?{$querystring}");
-        curl_setopt($ch, CURLOPT_USERPWD, "{$user_id}:{$user_password}");
-        curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-        curl_setopt($ch, CURLOPT_TIMEOUT, 10);
-
-        $json_result = curl_exec($ch);
-        curl_close($ch);
-
-        $result = json_decode($json_result, true);
-
-        return $result['is_valid'];
-    }
-
-    /**
-     * @param string[] $recipients
-     *
-     * @return string|false
-     */
-    private static function createRecipientVariables(array $recipients)
-    {
-        if (!is_array($recipients)) {
-            $recipients = [$recipients];
-        }
-        $recipient_vars = [];
-        foreach ($recipients as $recipient) {
-            $recipient_vars[$recipient] = [
-                'optout_link' => self::getEmailOptoutLinkByEmail($recipient),
-            ];
-        }
-
-        return json_encode($recipient_vars);
-    }
-
-    private static function getEmailOptoutLinkByEmail(string $email): string
-    {
-        $domain = \Config::$MAILGUN_SEND_DOMAIN;
-        $token = self::encryptedEmail($email);
-
-        return "https://$domain/account/marketing-agreement/email?token=$token";
-    }
-
-    /**
-     * @param string $email
-     *
-     * @return string|bool
-     */
-    private static function encryptedEmail(string $email)
-    {
-        return Encryption::encryptAESBase64($email, \Config::$MAILGUN_SECRET_KEY_OPT_OUT);
-    }
-
-    private static function generateMailgunAttachmentFormat(string $name, string $attachment): array
+    private function generateMailgunAttachmentFormat(string $name, string $attachment): array
     {
         return [
             'filePath' => $attachment,
-            'remoteName' => $name,
+            'filename' => $name,
         ];
+    }
+
+    private function assert(): void
+    {
+        if (empty($this->send_domain)) {
+            throw new \InvalidArgumentException('Not set send_domain. require them');
+        }
     }
 }